feat: Storm Zone (#3)
Adds Storm Zone

The Storm Zone only affects combats. In combat, a random player or monster will be given a Conduit Power. The owner of this power will be damaged for 10% of their max HP, but players only get damaged for 5% of their max HP. It's blockable damage and is not affected by other statuses (like Vulnerable).

At the start of each turn, a random card in hand is made damp. Damp cards retain and exhaust.

At the end of combat, there is a 50% chance once of the cards in the card reward is made Electric. Electric cards give one energy when played.

The zone darkens the players, and has a rain effect in the background. Visual lightning strikes in the background as well, lighting up the creatures in combat. This is purely visual. The owner of the conduit power will have an effect below them during the turn that applies more rapidly once the turn ends to signify lighting is about to strike. A light rain ambiance and thunder sound plays while in the zone, but it can be toggled via the ambient setting or managed with the SFX volume.

The Damp/Electric cards have visual effects applied to them and when the player uses an electric card, they get the shader applied to them for a bit too.

One thing to maybe TODO is create a config for turning off the various shaders.